Stroke Width Default Set Correctly to 1.0
Default stroke width sets to 1.0 and is, by default, subject to scaling factors. While the factors would affect a stroke previously the lack of a default required rather than None allowed the stroke to set later causing some strokes to appear the wrong size. This also removes invalidated equals based on stroke width. Equals is a property of stroke, fill, and geometry now.